% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNA.R
\name{CalculateNA}
\alias{CalculateNA}
\title{Calculate percentage and number of missing values for each variables}
\usage{
CalculateNA(df)
}
\arguments{
\item{df}{This is a dataframe}
}
\value{
The percentage and number of NAs for each variable
}
\description{
Calculate percentage and number of missing values for each variables
}
