% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare.R
\docType{data}
\name{healthcare}
\alias{healthcare}
\title{German Health Care Usage Dataset: 2004-2020}
\format{
27327 X 43 dataframe
}
\usage{
healthcare
}
\description{
This Panel data used in ECON 430 are from the German Health Care Usage Dataset, consist of 7,293 individuals across varying numbers of periods with a total of 27,326 observations.
}
\keyword{datasets}
