% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\docType{data}
\name{school}
\alias{school}
\title{Web search Interest Score in UCB and Stanford: 2004-2020}
\format{
204 X 4 dataframe
\describe{
\item{column1}{Year}
\item{column2}{Month}
\item{column3}{Interest Score in UCB}
\item{column3}{Interest Score in Stanford}
}
}
\usage{
school
}
\description{
A Time series data example for MAE project
}
\keyword{datasets}
