% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plus.R
\name{summary_plus}
\alias{summary_plus}
\title{Advanced Summary for A Data Frame}
\usage{
summary_plus(df)
}
\arguments{
\item{df}{This should be a date frame}
}
\value{
Returns Size, Numbers of missing value, Rows, Columns of the data frame.
    Returns missing value number and percentage and descriptive statistics for each variable.
}
\description{
An integrated solution for descriptive analysis.
    This function provides three main kinds of results:
    1. Size, Numbers of missing value, Rows, Columns of the data frame
    2. Numbers and percentage of missing value for each variable.
    3. Other descriptive statistics for each variable.
}
