% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucla.R
\docType{data}
\name{ucla}
\alias{ucla}
\title{Web search Interest Score in UCLA and USC: 2004-2020}
\format{
203 X 4 dataframe
\describe{
\item{column1}{Year}
\item{column2}{Month}
\item{column3}{Interest Score in UCLA}
\item{column3}{Interest Score in USC}
}
}
\usage{
ucla
}
\description{
A Time series data example for MAE project collected from Google Trends
}
\keyword{datasets}
